enabled security on  endpoints

+++ wifi4eu-portal/wifi4eu-portal-web-rest/src/main/java/wifi4eu/wifi4eu/web/rest/HelpdeskIssueResource.java
@@ -6,13 +6,20 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import wifi4eu.wifi4eu.common.dto.model.HelpdeskIssueDTO;
+import wifi4eu.wifi4eu.common.dto.model.UserDTO;
 import wifi4eu.wifi4eu.common.dto.rest.ErrorDTO;
 import wifi4eu.wifi4eu.common.dto.rest.ResponseDTO;
+import wifi4eu.wifi4eu.common.ecas.UserHolder;
+import wifi4eu.wifi4eu.common.security.UserContext;
 import wifi4eu.wifi4eu.service.helpdesk.HelpdeskService;
+import wifi4eu.wifi4eu.service.user.UserService;
 
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.List;
 
 @CrossOrigin(origins = "*")
@@ -23,20 +30,45 @@ public class HelpdeskIssueResource {
     @Autowired
     private HelpdeskService helpdeskService;
 
+    @Autowired
+    private UserService userService;
+
     Logger _log = LoggerFactory.getLogger(CallResource.class);
 
     @ApiOperation(value = "Get all the helpdesk issues")
     @RequestMapping(method = RequestMethod.GET, produces = "application/json")
     @ResponseBody
-    public List<HelpdeskIssueDTO> allHelpdeskIssues() {
+    public List<HelpdeskIssueDTO> allHelpdeskIssues(HttpServletResponse response) throws IOException {
+        try{
+            UserDTO userDTO = userService.getUserByUserContext(UserHolder.getUser());
+            if(userDTO.getType() != 5){
+                throw new AccessDeniedException(HttpStatus.NOT_FOUND.getReasonPhrase());
+            }
+        } catch (AccessDeniedException ade){
+            response.sendError(HttpStatus.NOT_FOUND.value());
+        } catch (Exception e){
+            response.sendError(HttpStatus.INTERNAL_SERVER_ERROR.value());
+        }
         return helpdeskService.getAllHelpdeskIssues();
     }
 
     @ApiOperation(value = "Get helpdesk issue by specific id")
     @RequestMapping(value = "/{issueId}", method = RequestMethod.GET, produces = "application/json")
     @ResponseBody
-    public HelpdeskIssueDTO getHelpdeskIssueById(@PathVariable("issueId") final Integer issueId) {
+    public HelpdeskIssueDTO getHelpdeskIssueById(@PathVariable("issueId") final Integer issueId, HttpServletResponse response) {
         _log.info("getHelpdeskIssueById: " + issueId);
+
+        try {
+            UserDTO userDTO = userService.getUserByUserContext(UserHolder.getUser());
+
+            if(userDTO.getType() != 5){
+                response.sendError(HttpStatus.NOT_FOUND.value());
+                return null;
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
         return helpdeskService.getHelpdeskIssueById(issueId);
     }
 
@@ -61,12 +93,22 @@ public class HelpdeskIssueResource {
     @ApiOperation(value = "Delete helpdesk by specific id")
     @RequestMapping(method = RequestMethod.DELETE)
     @ResponseBody
-    public ResponseDTO deleteHelpdeskIssue(@RequestBody final Integer issueId) {
+    public ResponseDTO deleteHelpdeskIssue(@RequestBody final Integer issueId, HttpServletResponse response) throws IOException {
         try {
             _log.info("deleteHelpdeskIssue: " + issueId);
+            UserDTO userDTO = userService.getUserByUserContext(UserHolder.getUser());
+            if(userDTO.getType() != 5){
+                response.sendError(HttpStatus.NOT_FOUND.value());
+                throw new AccessDeniedException(HttpStatus.NOT_FOUND.getReasonPhrase());
+            }
             HelpdeskIssueDTO resHelpdeskIssue = helpdeskService.deleteHelpdeskIssue(issueId);
             return new ResponseDTO(true, resHelpdeskIssue, null);
-        } catch (Exception e) {
+        }
+        catch (AccessDeniedException ade) {
+            response.sendError(HttpStatus.NOT_FOUND.value());
+            return new ResponseDTO(false, null, new ErrorDTO(HttpStatus.NOT_FOUND.value(), ade.getMessage()));
+        }
+        catch (Exception e) {
             if (_log.isErrorEnabled()) {
                 _log.error("Error on 'deleteHelpdeskIssue' operation.", e);
             }
